refactoring

+++ src/test/java/agh/po/ewolucja/AnimalTest.java
@@ -12,15 +12,15 @@ public class AnimalTest {
     @BeforeEach
     public void prepare(){
         map = new JungleMap(new Vector2d(20,20), new Vector2d(3,3));
-        a = new Animal(map, new Vector2d(10, 10), 100);
-        a2 = new Animal(map, new Vector2d(0,0), 100);
+        a = new Animal(map, new Vector2d(10, 10), 100.0);
+        a2 = new Animal(map, new Vector2d(0,0), 100.0);
         map.place(a);
         map.place(a2);
     }
 
     @Test
     public void moveToTest(){
-        int en = a.getEnergy();
+        Double en = a.getEnergy();
         Vector2d free = map.getFreeSpot(a.getPosition());
         a.moveTo(free);
         assertEquals(a.getPosition(), free);
@@ -29,17 +29,17 @@ public class AnimalTest {
 
     @Test
     public void eatTest(){
-        int e = a.getEnergy();
+        Double e = a.getEnergy();
         Vector2d free = map.getFreeSpot(a.getPosition());
-        Grass g = new Grass(free, 100);
+        Grass g = new Grass(free, 100.0);
         a.eat(g);
         assertEquals(a.getEnergy().intValue(), e+100);
     }
 
     @Test
     public void addEnergyTest(){
-        int e = a.getEnergy();
-        a.addEnergy(123);
+        Double e = a.getEnergy();
+        a.addEnergy(123.0);
         assertEquals(a.getEnergy().intValue(), e + 123);
     }
